% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipreg.R
\name{cred.subgroup}
\alias{cred.subgroup}
\title{Constructs a credible subset pair}
\usage{
cred.subgroup(
  params,
  design = NULL,
  FUN = function(x, params) {     params \%*\% t(x) },
  cred.level = 0.95,
  threshold = 0,
  method = c("asymptotic", "quantile"),
  step.down = TRUE,
  sides = c("both", "exclusive", "inclusive"),
  est.FUN = mean,
  var.FUN = sd,
  point.estimate = NULL,
  track = numeric(0),
  verbose = FALSE
)
}
\arguments{
\item{design}{(Optional) A numeric matrix whose rows are covariate points over which the band is to be constructed.}

\item{cred.level}{the credible level.}

\item{threshold}{the value of FUN above which a covariate is included in the target subset.}

\item{method}{Either "asymptotic" (default) or "quantile"; see details.}

\item{step.down}{Logical (default TRUE); should the step-down procedure be used?}

\item{sides}{One of "both" (default), "exclusive", or "inclusive". Which bounds should be constructed?}

\item{est.FUN}{The function used to produce estimates of the regression surface. Default is mean.}

\item{var.FUN}{The function used to quantify the variability of the regression surface posterior. Default is sd.}

\item{point.estimate}{If not null, replaces the mean and sets the reference around which the standard error is computed. Useful for bootstrapping methods. Treated as a row of the params matrix.}

\item{track}{A numeric vector of indices indicating which rows (default none) of the design matrix should have the sample of the corresponding FUN(x, params) returned.}

\item{verbose}{Logical (default FALSE); print progress?}

\item{param}{A numeric matrix whose rows are draws from the posterior distribution of either the regression surface or the parameter vector.}

\item{FUN(Optional)}{a function of the form function(x, params) that takes a row of design and the entire params matrix and returns a vector of the same length of x representing the regression surface.}
}
\value{
exclusive A logical vector indicating membership in the exclusive credible subset.

inclusive A logical vector indicating membership in the inclusive credible subset.

cred.level As provided.

threshold As provided.

method As provided.

step.down As provided.

sides As provided.

est Posterior estimate of the regression surface.

est.FUN As provided.

var Summary of posterior variability of the regression surface.

var.FUN As provided.

W An estimate of the extremal errors.

W.crit The critical quantile of W.

trace The posterior samples of the regression surface indicated by the track argument.
}
\description{
cred.subgroup returns a credible subset pair over a finite set of covariate points given either a sample from the posterior of the regression surface or a function FUN(x, params) and a sample from the posterior of the parameters.
}
\details{
If design is NULL (default), it is taken to be the identity matrix of dimension ncol(params), so that the rows of params are treated as draws from the posterior FUN(x, params).The 'asymptotic' method assumes that the marginal posteriors of the FUN(x, params) are asymptotically normal and is usually significantly faster and less memory-intensive than the 'quantile' method, which makes no such assumption.
}
