% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipreg.R
\docType{data}
\name{zipModel}
\alias{zipModel}
\title{specifies the prior and likelihood}
\format{
An object of class \code{character} of length 1.
}
\usage{
zipModel
}
\description{
specifies the prior and likelihood
}
\keyword{datasets}
