% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipreg.R
\name{zippte}
\alias{zippte}
\title{Estimate the personalized treatment effect}
\usage{
zippte(e.trt, e.ctrl)
}
\arguments{
\item{e.trt}{expected value for the outcome of the treatment}

\item{e.ctrl}{expected value for the outcome of the control}
}
\value{
personalized treatment effect(PTE)
}
\description{
Takes the expected value of treatment and control
}
