% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipreg.R
\name{zipreg}
\alias{zipreg}
\title{Estimate the coefficients of the ZIP model
Takes in response variable,prognostic and predictive covariates of both Poisson and Zero component}
\usage{
zipreg(y, trt, prog.z, prog.p, pred.p, nIter, offset.variable)
}
\arguments{
\item{offset.variable}{offset variable}

\item{Prog.z}{prognostic covariate of the logistic regression (zero component)}

\item{Pred.z}{predictive covariate of logistic regression (zero component)}

\item{Prog.p}{prognostic covariate of the Poisson regression (Poisson component)}

\item{Pred.p}{predictive covariate of Poisson regression (Poisson component)}

\item{Y}{count response variable}

\item{Trt}{treatment covariate}
}
\value{
coefficients of the Logistic and Poisson regression model
}
\description{
Estimate the coefficients of the ZIP model
Takes in response variable,prognostic and predictive covariates of both Poisson and Zero component
}
